Allow bootstrap build script to work on Mac OS X.
The -R option to cp, which is implied by -a, has behavior that varies
depending on whether the filenames to be copied have trailing slashes
or not. This does not appear to make a difference in the GNU coretools
version of cp, so I updated the script to allow bootstrap builds on
Mac OS X while maintaining compatibility with GNU coreutils.